Update README.md
Originally I posted the u2f token with keys I don't use to intuitively show the intended use.  Due to the number of people that assume the keys could be to my house/office, I've retaken the photo.  This is best practice anyways, regardless of keys being fake or not.