Move the C4 label out from under the fiducial.
Discovered the label was being destroyed since it was in the fiducial
keep-out zone. Needed to move the GND trace in order to make room next
to the cap.

Unfortunately, we don't have a panelization script so I simply tweaked
the panel by hand, which isn't ideal. Post-DRC fills look fine.