U2F HID custom/mandatory command distinction error bugfix
Continuation U2F HID packets with a sequence number equal to a custom command code are mistakenly processed as a custom command which leads to corrupted ping response and unwanted custom command execution.
The current bugfix corrects this issue, the firmware works well now with any size of ping commands.